requirements.txt
openrgb-python==0.3.3